+++ quil-rs/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.34.0-rc.2 (2025-11-07)
+
+### Breaking Changes
+
+- defgate as sequence
+
+## 0.34.0-rc.1 (2025-11-06)
+
+### Breaking Changes
+
+- defgate as sequence
+
+## 0.34.0-rc.0 (2025-11-05)
+
+### Breaking Changes
+
+- defgate as sequence
+
 ## 0.33.0 (2025-10-17)
 
 ### Breaking Changes